requirements.txt
# Dependências do Projeto
pdfplumber
pandas
pytesseract
pdf2image
pillow
python-dotenv

# Dependências da Documentação (Para o Netlify)
mkdocs
mkdocs-material
mkdocstrings[python]
mkdocs-encryptcontent-plugin